Upgraded the useraddsystem!
Upgrade complete! The interface is now much more userfriendly, and it is way clearer now who you've added. Instead of a long scrollable list we now have a searchbar, where you can enter the name or part of the name of the person you want to add, and it shows everyone who has that part in their name. You can instantly add them, and also immediatly delete them again, if that is what you want to do...